fix: remove unnecessary cast and pass clippy check

+++ stellargrant-contracts/contracts/stellar-grants/src/lib.rs
@@ -756,7 +756,7 @@ impl StellarGrantsContract {
                 &env,
                 grant_id,
                 milestone_idx,
-                milestone.approvals as u32,
+                milestone.approvals,
                 grant.quorum,
             );
